started moving search apis to /apis/search.js
