+++ src/js/models/v2/search/visualizations/geo/BaseSpendingByGeographyResult.js
@@ -3,8 +3,6 @@
  * Created by Andrea Blackwell 12/14/2023
  */
 
-import { formatMoneyWithPrecision } from 'helpers/moneyFormatter';
-
 const BaseSpendingByGeographyResult = {
     populate(data) {
         this.shape_code = data.shape_code;
@@ -12,12 +10,6 @@ const BaseSpendingByGeographyResult = {
         this.aggregated_amount = data.aggregated_amount || 0;
         this.population = data.population || 0;
         this.per_capita = data.per_capita || 0;
-    },
-    get aggregatedAmountFormatted() {
-        return formatMoneyWithPrecision(this.aggregated_amount, 0);
-    },
-    get perCapitaAggregated() {
-        return formatMoneyWithPrecision(this.per_capita, 0);
     }
 };
